fix: has badge issue
